moving waffle-http to waffle-mock.
