--- Source/JNA/waffle-http/.classpath
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry excluding="**/.svn|**/.svn/*" kind="src" path="src"/>
-	<classpathentry excluding="**/.svn|**/.svn/*" kind="src" path="src-test"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="/thirdparty/_lib/jar/javax.servlet/2.5/servlet-api-2.5.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
