--- Source/JNA/waffle-http/src-test/.gitkeep
@@ -1,1 +1,0 @@
-# keep
