+++ Source/JNA/waffle-mock/.classpath
@@ -6,5 +6,7 @@
 	<classpathentry combineaccessrules="false" kind="src" path="/waffle-api"/>
 	<classpathentry kind="lib" path="/thirdparty/net.java.dev.jna/3.4.2/jna.jar"/>
 	<classpathentry kind="lib" path="/thirdparty/net.java.dev.jna/3.4.2/platform.jar"/>
+	<classpathentry kind="lib" path="/thirdparty/_lib/jar/org.mockito/1.9.5-rc1/mockito-all-1.9.5-rc1.jar"/>
+	<classpathentry kind="lib" path="/thirdparty/_lib/jar/javax.servlet/3.0.1/javax.servlet-api-3.0.1.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
