+++ Source/JNA/waffle-mock/build.xml
@@ -3,9 +3,22 @@
 	<import file="${basedir}/../build/common.xml" />
 	<property name="project.jar" value="waffle-mock" />
 	<import file="${basedir}/../build/build.xml" />
+  <path id="javax.classpath">
+  	<!-- Either implementaion will work! -->
+  	<!--
+    <fileset dir="${thirdparty.dir}/_lib/jar/javax.servlet/2.5">
+      <include name="*.jar" />
+    </fileset>
+    -->
+    <fileset dir="${thirdparty.dir}/_lib/jar/javax.servlet/3.0.1">
+      <include name="*.jar" />
+    </fileset>
+  </path>
 	<path id="waffle-api.classpath" location="${common.build}/waffle-api.jar" />
 	<path id="project.classpath">
 		<path refid="waffle-api.classpath" />
+    <path refid="javax.classpath" />
+		<path refid="mockito.classpath" />
 	</path>
 	<path id="project.test.classpath" />
 </project>
