+++ Source/JNA/waffle-mock/src/waffle/mock/http/SimpleHttpRequest.java
@@ -11,27 +11,22 @@
  * Contributors:
  *     Application Security, Inc.
  *******************************************************************************/
-package waffle.http;
+package waffle.mock.http;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.security.Principal;
-import java.util.Enumeration;
 import java.util.HashMap;
-import java.util.Locale;
 import java.util.Map;
 
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletInputStream;
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletRequestWrapper;
 import javax.servlet.http.HttpSession;
+
+import org.mockito.Mockito;
 
 /**
  * @author dblock[at]dblock[dot]org
  */
-public class SimpleHttpRequest implements HttpServletRequest {
+public class SimpleHttpRequest extends HttpServletRequestWrapper {
 
 	private static int _remotePort_s = 0;
 
@@ -49,6 +44,7 @@ public class SimpleHttpRequest implements HttpServletRequest {
 	private Principal _principal = null;
 
 	public SimpleHttpRequest() {
+	  super(Mockito.mock(HttpServletRequest.class));
 		_remotePort = nextRemotePort();
 	}
 
@@ -173,210 +169,5 @@ public class SimpleHttpRequest implements HttpServletRequest {
 
 	public void setUserPrincipal(Principal principal) {
 		_principal = principal;
-	}
-
-	@Override
-	public Object getAttribute(String arg0) {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getAttributeNames() {
-		return null;
-	}
-
-	@Override
-	public String getCharacterEncoding() {
-		return null;
-	}
-
-	@Override
-	public String getContentType() {
-		return null;
-	}
-
-	@Override
-	public ServletInputStream getInputStream() throws IOException {
-		return null;
-	}
-
-	@Override
-	public String getLocalAddr() {
-		return null;
-	}
-
-	@Override
-	public String getLocalName() {
-		return null;
-	}
-
-	@Override
-	public int getLocalPort() {
-		return 0;
-	}
-
-	@Override
-	public Locale getLocale() {
-		return null;
-	}
-
-	@Override
-	public Enumeration<Locale> getLocales() {
-		return null;
-	}
-
-	@Override
-	public Map<String, String[]> getParameterMap() {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getParameterNames() {
-		return null;
-	}
-
-	@Override
-	public String[] getParameterValues(String arg0) {
-		return new String[0];
-	}
-
-	@Override
-	public String getProtocol() {
-		return null;
-	}
-
-	@Override
-	public BufferedReader getReader() throws IOException {
-		return null;
-	}
-
-	@Override
-	public String getRealPath(String arg0) {
-		return null;
-	}
-
-	@Override
-	public RequestDispatcher getRequestDispatcher(String arg0) {
-		return null;
-	}
-
-	@Override
-	public String getScheme() {
-		return null;
-	}
-
-	@Override
-	public String getServerName() {
-		return null;
-	}
-
-	@Override
-	public int getServerPort() {
-		return 0;
-	}
-
-	@Override
-	public boolean isSecure() {
-		return false;
-	}
-
-	@Override
-	public void removeAttribute(String arg0) {
-
-	}
-
-	@Override
-	public void setAttribute(String arg0, Object arg1) {
-
-	}
-
-	@Override
-	public void setCharacterEncoding(String arg0)
-			throws UnsupportedEncodingException {
-	}
-
-	@Override
-	public String getAuthType() {
-		return null;
-	}
-
-	@Override
-	public String getContextPath() {
-		return null;
-	}
-
-	@Override
-	public Cookie[] getCookies() {
-		return new Cookie[0];
-	}
-
-	@Override
-	public long getDateHeader(String arg0) {
-		return 0;
-	}
-
-	@Override
-	public Enumeration<String> getHeaderNames() {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getHeaders(String arg0) {
-		return null;
-	}
-
-	@Override
-	public int getIntHeader(String arg0) {
-		return 0;
-	}
-
-	@Override
-	public String getPathInfo() {
-		return null;
-	}
-
-	@Override
-	public String getPathTranslated() {
-		return null;
-	}
-
-	@Override
-	public StringBuffer getRequestURL() {
-		return null;
-	}
-
-	@Override
-	public String getRequestedSessionId() {
-		return null;
-	}
-
-	@Override
-	public String getServletPath() {
-		return null;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdFromCookie() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdFromURL() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdFromUrl() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdValid() {
-		return false;
-	}
-
-	@Override
-	public boolean isUserInRole(String arg0) {
-		return false;
 	}
 }
