+++ Source/JNA/waffle-servlet/src-test/waffle/util/AuthorizationHeaderTests.java
@@ -19,7 +19,7 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
-import waffle.http.SimpleHttpRequest;
+import waffle.mock.http.SimpleHttpRequest;
 
 /**
  * @author dblock[at]dblock[dot]org
