+++ Source/JNA/waffle-spring-security3/src-test/waffle/spring/NegotiateSecurityFilterTests.java
@@ -33,9 +33,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-import waffle.http.SimpleFilterChain;
-import waffle.http.SimpleHttpRequest;
-import waffle.http.SimpleHttpResponse;
+import waffle.mock.http.SimpleFilterChain;
+import waffle.mock.http.SimpleHttpRequest;
+import waffle.mock.http.SimpleHttpResponse;
 import waffle.servlet.spi.BasicSecurityFilterProvider;
 import waffle.servlet.spi.NegotiateSecurityFilterProvider;
 import waffle.servlet.spi.SecurityFilterProviderCollection;
@@ -112,7 +112,7 @@ public class NegotiateSecurityFilterTests {
 		assertEquals("ROLE_USER", authoritiesIterator.next().getAuthority());
 		assertEquals("ROLE_USERS", authoritiesIterator.next().getAuthority());
 		assertEquals("ROLE_EVERYONE", authoritiesIterator.next().getAuthority());
-		assertEquals(0, response.getHeaderNames().length);
+		assertEquals(0, response.getHeaderNamesSize());
 	}
 
 	@Test
